refactor: rename models to make them in consistency
