+++ src/lib/server/models/user.ts
@@ -1,7 +1,6 @@
+import { users } from '$lib/server/db/schema';
 import { eq } from 'drizzle-orm';
 import type { DrizzleD1Database } from 'drizzle-orm/d1';
-
-import { users } from '../db/schema';
 
 export async function getUserIdByName(db: DrizzleD1Database, name: string) {
 	return await db.select({ id: users.id }).from(users).where(eq(users.name, name)).get();
