+++ src/routes/api/clusters/[cluster_uuid]/+server.ts
@@ -1,6 +1,6 @@
 import { getDB } from '$lib/server/db';
-import { deleteCluster, getCluster, updateCluster } from '$lib/server/models/clusters';
-import { deleteNodes } from '$lib/server/models/nodes';
+import { deleteCluster, getCluster, updateCluster } from '$lib/server/models/cluster';
+import { deleteNodes } from '$lib/server/models/node';
 import type { ClusterDbUpdateInput, ClusterPublic } from '$lib/types/cluster';
 import type { D1Database } from '@cloudflare/workers-types';
 import type { RequestHandler } from '@sveltejs/kit';
