+++ src/routes/api/clusters/[cluster_uuid]/authority-map/+server.ts
@@ -1,5 +1,5 @@
 import { getDB } from '$lib/server/db';
-import { getNodes } from '$lib/server/models/nodes';
+import { getNodes } from '$lib/server/models/node';
 import type { Node } from '$lib/types/node';
 import type { D1Database } from '@cloudflare/workers-types';
 import { json } from '@sveltejs/kit';
