+++ src/routes/api/clusters/[cluster_uuid]/update-timestamp/+server.ts
@@ -1,5 +1,5 @@
 import { getDB } from '$lib/server/db';
-import { updateClusterTimestamp } from '$lib/server/models/clusters';
+import { updateClusterTimestamp } from '$lib/server/models/cluster';
 import type { D1Database } from '@cloudflare/workers-types';
 import type { RequestHandler } from '@sveltejs/kit';
 import { json } from '@sveltejs/kit';
